+++ r6-generator-maven-archetype/src/main/resources/archetype-resources/tests/testthat.R
@@ -1,5 +1,0 @@
-library(testthat)
-library(${rPackageName})
-library(tidyverse)
-
-test_check("${rPackageName}")