+++ r6-generator-runtime/src/main/java/uk/co/terminological/rjava/RMethod.java
@@ -21,14 +21,18 @@ import java.lang.annotation.Target;
 public @interface RMethod {
 
 	/**
-	 * Populate R examples
-	 * @return A list of R commands that will be executed as an example. 
+	 * Populate R examples. For non static methods this will be combined with @RClass(exampleSetup) annotation to
+	 * construct a complete example. For static methods (and constructors) the examples will be run as is. Use try({}) 
+	 * in examples you are not sure about or are developing. Please use single quotes only in R expressions (don't try escaping double quotes)
+	 * @return A list of R commands that will be executed as an example of this function.
 	 */
 	String[] examples() default {};
 	
 	/**
-	 * Populate R testthat tests
-	 * @return A list of R commands that will be executed as an test of this function. 
+	 * Populate R testthat tests. For non static methods this will be combined with @RClass(testSetup) annotation to
+	 * construct a complete test. For static methods (and constructors) the examples will be run as is. Use try({}) 
+	 * in examples you are not sure about or are developing. Please use single quotes only in R expressions (don't try escaping double quotes)
+	 * @return A list of R commands that will be executed as a test of this function. 
 	 */
 	String[] tests() default {};
 }